+++ src/main/java/wooteco/subway/controller/LineController.java
@@ -1,0 +1,83 @@
+package wooteco.subway.controller;
+
+import java.net.URI;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import wooteco.subway.domain.line.Line;
+import wooteco.subway.dto.line.LineRequest;
+import wooteco.subway.dto.line.LineResponse;
+import wooteco.subway.service.LineService;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.service.SectionService;
+import wooteco.subway.service.StationService;
+
+@RestController
+@RequestMapping("/lines")
+public class LineController {
+
+    private final LineService lineService;
+    private final SectionService sectionService;
+    private final StationService stationService;
+
+    @Autowired
+    public LineController(LineService lineService, SectionService sectionService,
+        StationService stationService) {
+        this.lineService = lineService;
+        this.sectionService = sectionService;
+        this.stationService = stationService;
+    }
+
+    @PostMapping
+    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
+        Line createdLine = lineService.createLine(new Line(lineRequest));
+        Section createdSection = new Section(createdLine.getId(), lineRequest);
+
+        stationService.showStation(lineRequest.getUpStationId());
+        stationService.showStation(lineRequest.getDownStationId());
+
+        sectionService.createSection(createdSection);
+
+        LineResponse lineResponse = LineResponse.from(createdLine);
+        return ResponseEntity.created(URI.create("/lines/" + createdLine.getId())).body(lineResponse);
+    }
+
+    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<LineResponse>> showLines() {
+        List<LineResponse> lineResponses = lineService.showLines().stream()
+            .map(LineResponse::from)
+            .collect(Collectors.toList());
+        return ResponseEntity.ok().body(lineResponses);
+    }
+
+    @GetMapping("{id}")
+    public ResponseEntity<LineResponse> showLine(@PathVariable long id) {
+        LineResponse lineResponse = LineResponse.from(lineService.showLine(id));
+        return ResponseEntity.ok().body(lineResponse);
+    }
+
+    @PutMapping("{id}")
+    public ResponseEntity<Void> updateLine(@RequestBody LineRequest lineRequest, @PathVariable long id) {
+        Line line = new Line(lineRequest);
+        lineService.updateLine(id, line);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("{id}")
+    public ResponseEntity<Void> deleteLine(@PathVariable long id) {
+        lineService.deleteLine(id);
+        return ResponseEntity.noContent().build();
+    }
+}
