+++ src/main/java/wooteco/subway/dto/station/StationRequest.java
@@ -1,0 +1,16 @@
+package wooteco.subway.dto.station;
+
+public class StationRequest {
+    private String name;
+
+    public StationRequest() {
+    }
+
+    public StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
