+++ src/main/java/wooteco/subway/dto/station/StationResponse.java
@@ -1,0 +1,45 @@
+package wooteco.subway.dto.station;
+
+import java.util.Objects;
+
+import wooteco.subway.domain.station.Station;
+
+public class StationResponse {
+    private Long id;
+    private String name;
+
+    public StationResponse() {
+    }
+
+    public StationResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public StationResponse(Station station) {
+        this(station.getId(), station.getName());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+        if (o == null || getClass() != o.getClass())
+            return false;
+        StationResponse that = (StationResponse)o;
+        return Objects.equals(id, that.id) && Objects.equals(name, that.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name);
+    }
+}
