--- src/main/java/wooteco/subway/station/controller/StationController.java
@@ -1,52 +1,0 @@
-package wooteco.subway.station.controller;
-
-import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import wooteco.subway.station.Station;
-import wooteco.subway.station.dto.StationRequest;
-import wooteco.subway.station.dto.StationResponse;
-import wooteco.subway.station.service.StationService;
-
-@RestController
-@RequestMapping("/stations")
-public class StationController {
-
-    private final StationService stationService;
-
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
-    }
-
-    @PostMapping
-    public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
-        Station station = new Station(stationRequest.getName());
-        StationResponse stationResponse = new StationResponse(stationService.createStation(station));
-        return ResponseEntity.created(URI.create("/stations/" + stationResponse.getId())).body(stationResponse);
-    }
-
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<StationResponse>> showStations() {
-        List<StationResponse> stations = stationService.showStations().stream()
-            .map(station -> new StationResponse(station.getId(), station.getName()))
-            .collect(Collectors.toList());
-        return ResponseEntity.ok().body(stations);
-    }
-
-    @DeleteMapping("{id}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
-        stationService.deleteStation(id);
-        return ResponseEntity.noContent().build();
-    }
-}
