--- src/main/java/wooteco/subway/station/dao/StationDao.java
@@ -1,59 +1,0 @@
-package wooteco.subway.station.dao;
-
-import java.sql.PreparedStatement;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import org.springframework.validation.annotation.Validated;
-
-import wooteco.subway.station.Station;
-
-@Validated
-@Repository
-public class StationDao {
-
-    private final JdbcTemplate jdbcTemplate;
-    private final RowMapper<Station> stationRowMapper = (resultSet, rowNum) -> new Station(
-        resultSet.getLong("id"),
-        resultSet.getString("name")
-    );
-
-    public StationDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public long save(Station station) {
-        String sql = "INSERT INTO STATION (name) values (?)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(con -> {
-            PreparedStatement ps = con.prepareStatement(sql, new String[] {"id"});
-            ps.setString(1, station.getName());
-            return ps;
-        }, keyHolder);
-
-        return Objects.requireNonNull(keyHolder.getKey()).longValue();
-    }
-
-    public List<Station> findAll() {
-        String sql = "SELECT * FROM STATION";
-        return jdbcTemplate.query(sql, stationRowMapper);
-    }
-
-    public int delete(long stationId) {
-        String sql = "DELETE FROM STATION WHERE id=?";
-        return jdbcTemplate.update(sql, stationId);
-    }
-
-    public Optional<Station> findById(Long id) {
-        String sql = "SELECT * FROM STATION WHERE id = ?";
-        List<Station> result = jdbcTemplate.query(sql, stationRowMapper, id);
-        return result.stream().findAny();
-    }
-}
